refactor: remove backwards-compat re-exports from headless-browser
