+++ assistant/src/__tests__/headless-browser-interactions.test.ts
@@ -77,7 +77,7 @@ import {
   executeBrowserSelectOption,
   executeBrowserSnapshot,
   executeBrowserType,
-} from "../tools/browser/headless-browser.js";
+} from "../tools/browser/browser-execution.js";
 import type { ToolContext } from "../tools/types.js";
 
 const ctx: ToolContext = {
